move functions to utils.py
